AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA

+++ database/migrations/2022_04_19_190658_create_productos_table.php
@@ -18,7 +18,7 @@ class CreateProductosTable extends Migration
             $table->string('nombre');
             $table->string('descripcion',500);
             $table->double('precio');
-            $table->double('descuento');
+            $table->integer('descuento');
             $table->string('estadoProducto');
             $table->integer('stock');
             $table->string('foto');
